Changes to be committed: 	modified:   AUTHORS Updated personal info.
	modified:   Makefile
Beginning work to clean up Makefile. The original makefile appeared
to be checking if the system was Windows NT. Since this is for GNU/
Linux, that was removed. Also all of the code is C, but the
Makefile was calling the C++ compiler to hide bool issues and
missing typedefs (which C++ is ok with but C is not).

	modified:   src/daemon.c
Added structs, added stdbool.h

	modified:   src/main.c
Added stdbool.h

	modified:   src/mbpfan.c
Added stdbool.h

	modified:   src/minunit.c
Added stdbool.h